Fix bug copying features from tf-idf vector
`keys` is a list of the form `['account', 'advisor', ..., 'whistle']`.
`tfvec` is a list of the form `[[0., 0., ..., 0.]]`. This code combines
these two lists into `clean_row`, a dict of the form `{ 'account': 0.,
'advisor': 0., ..., 'whistle': 0. }`. Previously, `len(tfvec)` was being
used in the range. Since `tfvec` is a list of length 1, it was only
copying the first feature, with the result being `{ 'account': 0. }`.
This change uses `len(tfvec[0])` in the range so that all features are
copied.